+++ app/models/grouping.rb
@@ -456,7 +456,7 @@ class Grouping < ApplicationRecord
       end
     end
 
-    result
+    raise I18n.t('repo.assignment_dir_creation_error', short_identifier: assignment.short_identifier) unless result
   end
 
   # Get the section for this group. If assignment restricts member of a groupe